+++ .github/workflows/unity.yml
@@ -26,7 +26,7 @@ jobs:
         with:
           build-options: ./.github/workflows/build-options.json
           group-by: 'unity-version'
-          job-name-prefix: 'Build and Test'
+          job-name-prefix: 'Build'
     outputs:
       jobs: ${{ steps.setup-jobs.outputs.jobs }}
   validate:
