+++ ElevenLabs/Packages/com.rest.elevenlabs/Runtime/Voices/CategoryTypes.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1e0c2b8d5503b5c4fa45a9d74e47757f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 5b71cbcaf078a8e44a5c96dcd24376d5, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
