+++ ElevenLabs/Packages/com.rest.elevenlabs/Runtime/Voices/FineTuningStateTypes.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 536206db21f50d74ca581e8f3eed0a25
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 5b71cbcaf078a8e44a5c96dcd24376d5, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
