+++ ElevenLabs/Packages/com.rest.elevenlabs/Runtime/Voices/SortDirections.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ee0e1abd04f86c644a343ed8405f1ed6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 5b71cbcaf078a8e44a5c96dcd24376d5, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
