+++ ElevenLabs/Packages/com.rest.elevenlabs/Runtime/Voices/VoiceList.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5f8161b0297c2d542b0ab4a0cea1fd8d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 5b71cbcaf078a8e44a5c96dcd24376d5, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
