+++ ElevenLabs/Packages/com.rest.elevenlabs/Runtime/Voices/VoiceQuery.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b0d5d162085c03e43a36a7dbde7c54a6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 5b71cbcaf078a8e44a5c96dcd24376d5, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
