+++ ElevenLabs/Packages/com.rest.elevenlabs/Runtime/Voices/VoiceTypes.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d0000daffb433b94792f1b59e193f5b2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 5b71cbcaf078a8e44a5c96dcd24376d5, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
